🚚 rename makeListMessage
